Move `Asset` to `files.py`
