+++ pyunity/core.py
@@ -52,12 +52,12 @@ and all have MeshRenderers:
 
 __all__ = ["Component", "GameObject", "SingleComponent",
            "Tag", "Transform", "ShowInInspector",
-           "HideInInspector", "addFields", "Asset", "SavesProjectID"]
+           "HideInInspector", "addFields", "SavesProjectID"]
 
 import inspect
 import os
 from .errors import PyUnityException, ComponentException
-from .values import Vector3, Quaternion, ABCMeta, abstractmethod
+from .values import Vector3, Quaternion
 from . import Logger
 
 class Tag:
@@ -122,11 +122,6 @@ class Tag:
 
 class SavesProjectID:
     pass
-
-class Asset(SavesProjectID, metaclass=ABCMeta):
-    @abstractmethod
-    def SaveAsset(self, ctx):
-        pass
 
 class GameObject(SavesProjectID):
     """
